+++ subsync/subsync.py
@@ -1,13 +1,8 @@
 #!/usr/bin/env python
-from __future__ import print_function
-from builtins import range
-
 import argparse
-import math
 import logging
 import sys
 
-import numpy as np
 from sklearn.pipeline import Pipeline
 
 from .aligners import FFTAligner, MaxScoreAligner
@@ -22,7 +17,7 @@ FRAME_RATE = 48000
 SAMPLE_RATE = 100
 
 
-def make_srt_speech_pipeline(encoding='infer'):
+def make_srt_speech_pipeline(encoding):
     return Pipeline([
         ('parse', SrtParser(encoding=encoding)),
         ('speech_extract', SubtitleSpeechTransformer(sample_rate=SAMPLE_RATE))
@@ -33,30 +28,39 @@ def main():
     parser = argparse.ArgumentParser(description='Synchronize subtitles with video.')
     parser.add_argument('-v', '--version', action='version',
                         version='%(prog)s {version}'.format(version=__version__))
-    parser.add_argument('reference')
-    parser.add_argument('-i', '--srtin', required=True)  # TODO: allow read from stdin
-    parser.add_argument('-o', '--srtout', default=None)
+    parser.add_argument('reference', help='Correct reference (video or srt) to which to sync input subtitles.')
+    parser.add_argument('-i', '--srtin', help='Input subtitles file (default=stdin).')
+    parser.add_argument('-o', '--srtout', help='Output subtitles file (default=stdout).')
+    parser.add_argument('--encoding', default='infer',
+                        help='What encoding to use for reading input subtitles.')
+    parser.add_argument('--output-encoding', default='same',
+                        help='What encoding to use for writing output subtitles (default=same as for input).')
+    parser.add_argument('--reference-encoding',
+                        help='What encoding to use for reading / writing reference subtitles (if applicable).')
     parser.add_argument('--vlc-mode', action='store_true', help=argparse.SUPPRESS)
     args = parser.parse_args()
     if args.vlc_mode:
         logger.setLevel(logging.CRITICAL)
     if args.reference.endswith('srt'):
-        reference_pipe = make_srt_speech_pipeline()
+        reference_pipe = make_srt_speech_pipeline(args.reference_encoding or 'infer')
     else:
+        if args.reference_encoding is not None:
+            logger.warning('Reference srt encoding specified, but reference was a video file')
         reference_pipe = Pipeline([
             ('speech_extract', VideoSpeechTransformer(sample_rate=SAMPLE_RATE,
                                                       frame_rate=FRAME_RATE,
                                                       vlc_mode=args.vlc_mode))
         ])
-    srtin_pipe = make_srt_speech_pipeline()
+    srtin_pipe = make_srt_speech_pipeline(args.encoding)
     logger.info('computing alignments...')
     offset_seconds = MaxScoreAligner(FFTAligner).fit_transform(
         srtin_pipe.fit_transform(args.srtin),
         reference_pipe.fit_transform(args.reference)
-    ) / 100.
+    ) / float(SAMPLE_RATE)
     logger.info('offset seconds: %.3f', offset_seconds)
     SrtOffseter(offset_seconds).fit_transform(
-        srtin_pipe.named_steps['parse'].subs_).write_file(args.srtout)
+        srtin_pipe.named_steps['parse'].subs_).set_encoding(
+        args.output_encoding).write_file(args.srtout)
     return 0